Maintenance.

+++ src/AssemblyCounters.hpp
@@ -74,7 +74,7 @@ namespace Tensors {
             {
                 // each thread does the accumulation on its chunk independently
                 const Int j_begin = (step*(thread  ) + (corr*(thread  ))/thread_count) * per_line;
-                const Int j_end   = Min(m, (step*(thread+Int(1)) + (corr*(thread+Int(1)))/thread_count) * per_line);
+                const Int j_end   = Min(m, static_cast<Int>( (step*(thread+Int(1)) + (corr*(thread+Int(1)))/thread_count) * per_line) );
                 
                 if( j_end > j_begin )
                 {
@@ -120,7 +120,7 @@ namespace Tensors {
                 const LInt correction = S[thread];
                 
                 const Int j_begin = (step*(thread  ) + (corr*(thread  ))/thread_count) * per_line;
-                const Int j_end   = Min(m, (step*(thread+1) + (corr*(thread+1))/thread_count) * per_line);
+                const Int j_end   = Min(m, static_cast<Int>( (step*(thread+1) + (corr*(thread+1))/thread_count) * per_line) );
 
                 
                 for( Int i = 0; i < thread_count; ++i )
